requirements.txt
crawl4ai>=0.3.0
anthropic>=0.40.0
pydantic>=2.0.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
rich>=13.0.0
pyyaml>=6.0.0
